requirements.txt
streamlit
scikit-learn
pandas
PyPDF2
python-docx